Convert the decision table into long or wide format

`pivot_decision_tbl_wider`, pivots the decision table into a wider form with one paper per row and each column as a decision. `pivot_decision_tbl_longer` pivots all the columns, except paper and model, to the long format.

pivot_decision_tbl_wider(df)

pivot_decision_tbl_longer(df)

Arguments

df

A decision_tbl object created by `as_decision_tbl`

Value

A data frame in wide format with decisions split across columns.

Examples

raw_df <- read.csv(system.file("papers.csv", package = "dossier")) |> tibble::as_tibble()
tbl_df <- as_decision_tbl(raw_df)
pivot_decision_tbl_wider(tbl_df)
#> # A tibble: 6 × 30
#>   paper       model                barometric_pressure_…¹ barometric_pressure_…²
#>   <chr>       <chr>                <chr>                  <chr>                 
#> 1 braga       generalized additiv… LOESS                  to minimize Akaike's …
#> 2 katsouyanni generalized additiv… NA                     NA                    
#> 3 ostro       Poisson regression   NA                     NA                    
#> 4 peel        Poisson generalized… NA                     NA                    
#> 5 schwartz    Poisson regression   NA                     NA                    
#> 6 zanobetti   generalized additiv… NA                     NA                    
#> # ℹ abbreviated names: ¹​barometric_pressure_parameter_method,
#> #   ²​barometric_pressure_parameter_reason
#> # ℹ 26 more variables: barometric_pressure_parameter_decision <chr>,
#> #   barometric_pressure_spatial_reason <chr>,
#> #   barometric_pressure_spatial_decision <chr>,
#> #   humidity_parameter_method <chr>, humidity_parameter_reason <chr>,
#> #   humidity_parameter_decision <chr>, humidity_spatial_reason <chr>, …
pivot_decision_tbl_longer(tbl_df)
#> # A tibble: 64 × 8
#>    paper variable            type      model     parameter decision reason    id
#>    <chr> <chr>               <chr>     <chr>     <chr>     <chr>    <chr>  <int>
#>  1 braga barometric_pressure parameter generali… smoothin… baromet… LOESS      1
#>  2 braga barometric_pressure parameter generali… smoothin… baromet… to mi…     2
#>  3 braga barometric_pressure spatial   generali… NA        baromet… to al…     3
#>  4 braga barometric_pressure spatial   generali… NA        baromet… chose…     4
#>  5 braga humidity            parameter generali… smoothin… humidit… LOESS      5
#>  6 braga humidity            parameter generali… smoothin… humidit… to mi…     6
#>  7 braga humidity            spatial   generali… NA        humidit… to al…     7
#>  8 braga humidity            spatial   generali… NA        humidit… chose…     8
#>  9 braga temperature         parameter generali… smoothin… tempera… LOESS      9
#> 10 braga temperature         parameter generali… smoothin… tempera… to mi…    10
#> # ℹ 54 more rows